move accuracy_floor from mixin to gate test
